Misc undefined behavior fixes
- instructions_shift:

`lo` is 16bit but shifted << 16, so it needs to be 32bit.

- m68k_io:

Shift by 16 needs a value large enough to hold it, so we need to make that an
uint32.

- instructions_transfer:

False positive, but adding the parenthesis will content the sanitizer.

+++ megado/m68k/instructions_arithmetic.c
@@ -299,7 +299,7 @@ uint8_t divs(Instruction* i, M68k* ctx)
     }
     else {
         int32_t quotient = b / a;
-        int16_t remainder = b % a;
+        uint32_t remainder = b % a;
         ctx->data_registers[i->dst->n] = remainder << 16 | (quotient & 0xFFFF);
 
         CARRY_SET(ctx, false);
